% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{unif_random}
\alias{unif_random}
\title{Generate uniformly random numbers}
\usage{
unif_random(n, min = 0, max = 1)
}
\arguments{
\item{n}{A positive integer specifying the number of values to generate.}

\item{min}{The lower limit of the distribution.}

\item{max}{The upper limit of the distribution.}
}
\value{
A uniformly random number.
}
\description{
Generate uniformly random numbers
}
\references{
Steele, Guy; Vigna, Sebastiano (15 January 2020). "Computationally easy, spectrally good multipliers for congruential pseudorandom number generators". \href{https://arxiv.org/abs/2001.05304}{arXiv: 2001.05304}
}
